Back to explicit subquery renderer
